Second commit for #11. See commit body for description.
This commit refactors the previous implementation, and makes it optional to emit an extra event after checkout.

The commit also includes a couple of refactors.

1. I've created a series of very simple message objects that encapsulates the presentation and creation of the JSON
message bodies.

2. I've extracted all utility methods into static methods in classes so that i can re-use functionality in both SCMListener and BuildListener.

That's it. Please review and suggest.
