+++ src/main/java/org/datadog/jenkins/plugins/datadog/BuildFinishedEventImpl.java
@@ -1,0 +1,65 @@
+package org.datadog.jenkins.plugins.datadog;
+
+import java.util.HashMap;
+import net.sf.json.JSONObject;
+
+/**
+ *
+ * @author Mads
+ */
+public class BuildFinishedEventImpl implements DataDogEvent {
+
+  private JSONObject builddata;
+  private HashMap<String,String> tags;
+
+  public BuildFinishedEventImpl(JSONObject buildData, HashMap<String,String> tags)  {
+    this.builddata = buildData;
+    this.tags = tags;
+  }
+
+  //Creates the raw json payload for this event.
+  @Override
+  public JSONObject createPayload() {
+    JSONObject payload = new JSONObject();
+    // Add event_type to assist in roll-ups
+    payload.put("event_type", "build result"); // string
+    String hostname = DataDogUtilities.nullSafeGetString(builddata, "hostname");
+    String number = DataDogUtilities.nullSafeGetString(builddata, "number");
+    String buildurl = DataDogUtilities.nullSafeGetString(builddata, "buildurl");
+    String job = DataDogUtilities.nullSafeGetString(builddata, "job");
+    long timestamp = builddata.getLong("timestamp");
+    String message = "";
+
+    // Build title
+    StringBuilder title = new StringBuilder();
+    title.append(job).append(" build #").append(number);
+    if ("SUCCESS".equals(builddata.get("result"))) {
+      title.append(" succeeded");
+      payload.put("alert_type", "success");
+      message = "%%% \n [See results for build #" + number + "](" + buildurl + ") ";
+    } else if (builddata.get("result") != null) {
+      title.append(" failed");
+      payload.put("alert_type", "failure");
+      message = "%%% \n [See results for build #" + number + "](" + buildurl + ") ";
+    }
+    title.append(" on ").append(hostname);
+    // Add duration
+    if (builddata.get("duration") != null) {
+      message = message + DataDogUtilities.durationToString(builddata.getDouble("duration"));
+    }
+
+    // Close markdown
+    message = message + " \n %%%";
+
+    // Build payload
+    payload.put("title", title.toString());
+    payload.put("text", message);
+    payload.put("date_happened", timestamp);
+    payload.put("host", hostname);
+    payload.put("result", builddata.get("result"));
+    payload.put("tags", DataDogUtilities.assembleTags(builddata, tags));
+    payload.put("aggregation_key", job);
+
+    return payload;
+  }
+}
