+++ src/main/java/org/datadog/jenkins/plugins/datadog/DatadogBuildListener.java
@@ -4,7 +4,6 @@ import static hudson.Util.fixEmptyAndTrim;
 
 import hudson.EnvVars;
 import hudson.Extension;
-import hudson.ProxyConfiguration;
 import hudson.model.AbstractProject;
 import hudson.model.Describable;
 import hudson.model.Descriptor;
@@ -13,7 +12,6 @@ import hudson.model.TaskListener;
 import hudson.model.listeners.RunListener;
 import hudson.util.FormValidation;
 import hudson.util.Secret;
-import jenkins.model.Jenkins;
 
 import net.sf.json.JSONArray;
 import net.sf.json.JSONObject;
@@ -23,20 +21,12 @@ import org.kohsuke.stapler.QueryParameter;
 import org.kohsuke.stapler.StaplerRequest;
 
 import java.io.BufferedReader;
-import java.io.DataOutputStream;
 import java.io.IOException;
-import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.net.HttpURLConnection;
-import java.net.Inet4Address;
-import java.net.Proxy;
 import java.net.URL;
-import java.net.UnknownHostException;
-import java.util.Arrays;
 import java.util.HashMap;
 import java.util.logging.Logger;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 import javax.annotation.Nonnull;
 import javax.servlet.ServletException;
@@ -98,14 +88,14 @@ public class DatadogBuildListener extends RunListener<Run>
     String jobName = run.getParent().getDisplayName();
     HashMap<String,String> tags = new HashMap<String,String>();
     // Process only if job is NOT in blacklist
-    if ( isJobTracked(jobName) ) {
-      logger.fine("Started build!");
+    if ( DataDogUtilities.isJobTracked(jobName) ) {
+      logger.fine("Started build! in onStarted()");
 
       // Grab environment variables
       EnvVars envVars = null;
       try {
         envVars = run.getEnvironment(listener);
-        tags = parseTagList(run, listener);
+        tags = DataDogUtilities.parseTagList(run, listener);
       } catch (IOException e) {
         logger.severe(e.getMessage());
       } catch (InterruptedException e) {
@@ -114,7 +104,7 @@ public class DatadogBuildListener extends RunListener<Run>
 
       // Gather pre-build metadata
       JSONObject builddata = new JSONObject();
-      builddata.put("hostname", getHostname(envVars)); // string
+      builddata.put("hostname", DataDogUtilities.getHostname(envVars)); // string
       builddata.put("job", jobName); // string
       builddata.put("number", run.number); // int
       builddata.put("result", null); // null
@@ -123,9 +113,9 @@ public class DatadogBuildListener extends RunListener<Run>
       long starttime = run.getStartTimeInMillis() / DatadogBuildListener.THOUSAND_LONG; // ms to s
       builddata.put("timestamp", starttime); // string
 
-      // Add event_type to assist in roll-ups
-      builddata.put("event_type", "build start"); // string
-      event(builddata, tags);
+      BuildStartedEventImpl evt = new BuildStartedEventImpl(builddata, tags);
+
+      DataDogHttpRequests.sendEvent(evt);
     }
   }
 
@@ -141,25 +131,22 @@ public class DatadogBuildListener extends RunListener<Run>
   public final void onCompleted(final Run run, @Nonnull final TaskListener listener) {
     final String jobName = run.getParent().getDisplayName();
     // Process only if job in NOT in blacklist
-    if ( isJobTracked(jobName) ) {
+    if ( DataDogUtilities.isJobTracked(jobName) ) {
       logger.fine("Completed build!");
 
       // Collect Data
       JSONObject builddata = gatherBuildMetadata(run, listener);
       HashMap<String,String> extraTags = new HashMap<String, String>();
       try {
-        extraTags = parseTagList(run, listener);
+        extraTags = DataDogUtilities.parseTagList(run, listener);
       } catch (IOException ex) {
         logger.severe(ex.getMessage());
       } catch (InterruptedException ex) {
         logger.severe(ex.getMessage());
       }
 
-      // Add event_type to assist in roll-ups
-      builddata.put("event_type", "build result"); // string
-
-      // Report Data
-      event(builddata, extraTags);
+      DataDogEvent evt = new BuildFinishedEventImpl(builddata, extraTags);
+      DataDogHttpRequests.sendEvent(evt);
       gauge("jenkins.job.duration", builddata, "duration", extraTags);
       if ( "SUCCESS".equals(builddata.get("result")) ) {
         serviceCheck("jenkins.job.status", DatadogBuildListener.OK, builddata, extraTags);
@@ -200,7 +187,7 @@ public class DatadogBuildListener extends RunListener<Run>
     builddata.put("result", run.getResult().toString()); // string
     builddata.put("number", run.number); // int
     builddata.put("job", run.getParent().getDisplayName()); // string
-    builddata.put("hostname", getHostname(envVars)); // string
+    builddata.put("hostname", DataDogUtilities.getHostname(envVars)); // string
     builddata.put("buildurl", envVars.get("BUILD_URL")); // string
     builddata.put("node", envVars.get("NODE_NAME")); // string
 
@@ -212,153 +199,6 @@ public class DatadogBuildListener extends RunListener<Run>
     return builddata;
   }
 
-  /**
-   * Assembles a {@link JSONArray} from metadata available in the
-   * {@link JSONObject} builddata. Returns a {@link JSONArray} with the set
-   * of tags.
-   *
-   * @param builddata - A JSONObject containing a builds metadata.
-   * @param extra - A list of tags, that are contributed via {@link DataDogJobProperty}.
-   * @return a JSONArray containing a specific subset of tags retrieved from a builds metadata.
-   */
-  private JSONArray assembleTags(final JSONObject builddata, final HashMap<String,String> extra) {
-    JSONArray tags = new JSONArray();
-
-    tags.add("job:" + builddata.get("job"));
-    if ( (builddata.get("node") != null) && getDescriptor().getTagNode() ) {
-      tags.add("node:" + builddata.get("node"));
-    }
-
-    if ( builddata.get("result") != null ) {
-      tags.add("result:" + builddata.get("result"));
-    }
-
-    if ( builddata.get("branch") != null && !extra.containsKey("branch") ) {
-      tags.add("branch:" + builddata.get("branch"));
-    }
-
-    //Add the extra tags here
-    for(String key : extra.keySet()) {
-      tags.add(String.format("%s:%s", key, extra.get(key)));
-      logger.info(String.format("Emitted tag %s:%s", key, extra.get(key)));
-    }
-
-    return tags;
-  }
-
-  /**
-   * This method parses the contents of the configured DataDog tags. If they are present.
-   * Takes the current build as a parameter. And returns the expanded tags and their
-   * values in a HashMap.
-   *
-   * Always returns a HashMap, that can be empty, if no tagging is configured.
-   *
-   * @param run - Current build
-   * @param listener - Current listener
-   * @return A {@link HashMap} containing the key,value pairs of tags. Never null.
-   * @throws IOException
-   * @throws InterruptedException
-   */
-  @Nonnull
-  private HashMap<String,String> parseTagList(Run run, TaskListener listener) throws IOException,
-          InterruptedException {
-    HashMap<String,String> map = new HashMap<String, String>();
-
-    DataDogJobProperty property = (DataDogJobProperty)run.getParent()
-            .getProperty(DataDogJobProperty.class);
-    String prop = property.getTagProperties();
-
-    if( !property.isTagFileEmpty() ) {
-      String dataFromFile = property.readTagFile(run);
-      if(dataFromFile != null) {
-        for(String tag : dataFromFile.split("\\r?\\n")) {
-          String[] expanded = run.getEnvironment(listener).expand(tag).split("=");
-          if( expanded.length > 1 ) {
-            map.put(expanded[0], expanded[1]);
-            logger.fine(String.format("Emitted tag %s:%s", expanded[0], expanded[1]));
-          } else {
-            logger.fine(String.format("Ignoring the tag %s. It is empty.", tag));
-          }
-        }
-      }
-    }
-
-    if( !property.isTagPropertiesEmpty() ) {
-      for(String tag : prop.split("\\r?\\n")) {
-        String[] expanded = run.getEnvironment(listener).expand(tag).split("=");
-        if( expanded.length > 1 ) {
-          map.put(expanded[0], expanded[1]);
-          logger.fine(String.format("Emitted tag %s:%s", expanded[0], expanded[1]));
-        } else {
-          logger.fine(String.format("Ignoring the tag %s. It is empty.", tag));
-        }
-      }
-    }
-
-    return map;
-  }
-
-  /**
-   * Posts a given {@link JSONObject} payload to the DataDog API, using the
-   * user configured apiKey.
-   *
-   * @param payload - A JSONObject containing a specific subset of a builds metadata.
-   * @param type - A String containing the URL subpath pertaining to the type of API post required.
-   * @return a boolean to signify the success or failure of the HTTP POST request.
-   * @throws IOException
-   */
-  public final Boolean post(final JSONObject payload, final String type) throws IOException {
-    String urlParameters = "?api_key=" + getDescriptor().getApiKey();
-    HttpURLConnection conn = null;
-
-    try {
-      // Make request
-      conn = DatadogBuildListener.getHttpURLConnection(new URL(DatadogBuildListener.BASEURL + type
-                                                               + urlParameters));
-      conn.setRequestMethod("POST");
-      conn.setRequestProperty("Content-Type", "application/json");
-      conn.setUseCaches(false);
-      conn.setDoInput(true);
-      conn.setDoOutput(true);
-
-      // Send request
-      DataOutputStream wr = new DataOutputStream( conn.getOutputStream() );
-      wr.writeBytes( payload.toString() );
-      wr.flush();
-      wr.close();
-
-      // Get response
-      BufferedReader rd = new BufferedReader( new InputStreamReader( conn.getInputStream() ) );
-      StringBuilder result = new StringBuilder();
-      String line;
-      while ( (line = rd.readLine()) != null ) {
-        result.append(line);
-      }
-      rd.close();
-      JSONObject json = (JSONObject) JSONSerializer.toJSON( result.toString() );
-      if ( "ok".equals(json.getString("status")) ) {
-        logger.finer(String.format("API call of type '%s' was sent successfully!", type));
-        logger.finer(String.format("Payload: %s", payload));
-        return true;
-      } else {
-        logger.fine(String.format("API call of type '%s' failed!", type));
-        logger.fine(String.format("Payload: %s", payload));
-        return false;
-      }
-    } catch (Exception e) {
-      if ( conn.getResponseCode() == DatadogBuildListener.HTTP_FORBIDDEN ) {
-        logger.severe("Hmmm, your API key may be invalid. We received a 403 error.");
-      } else {
-        logger.severe(String.format("Client error: %s", e));
-      }
-      return false;
-    } finally {
-      if (conn != null) {
-        conn.disconnect();
-      }
-      return true;
-    }
-  }
 
   /**
    * Sends a metric to the Datadog API, including the gauge name, and value.
@@ -371,7 +211,7 @@ public class DatadogBuildListener extends RunListener<Run>
    */
   public final void gauge(final String metricName, final JSONObject builddata,
                           final String key, final HashMap<String,String> extraTags) {
-    String builddataKey = nullSafeGetString(builddata, key);
+    String builddataKey = DataDogUtilities.nullSafeGetString(builddata, key);
     logger.fine(String.format("Sending metric '%s' with value %s", metricName, builddataKey));
 
     // Setup data point, of type [<unix_timestamp>, <value>]
@@ -389,7 +229,7 @@ public class DatadogBuildListener extends RunListener<Run>
     metric.put("points", points);
     metric.put("type", "gauge");
     metric.put("host", builddata.get("hostname"));
-    metric.put("tags", assembleTags(builddata, extraTags));
+    metric.put("tags", DataDogUtilities.assembleTags(builddata, extraTags));
 
     // Place metric as item of series list
     JSONArray series = new JSONArray();
@@ -402,7 +242,7 @@ public class DatadogBuildListener extends RunListener<Run>
     logger.fine(String.format("Resulting payload: %s", payload.toString() ));
 
     try {
-      post(payload, DatadogBuildListener.METRIC);
+      DataDogHttpRequests.post(payload, METRIC);
     } catch (Exception e) {
       logger.severe(e.toString());
     }
@@ -427,265 +267,13 @@ public class DatadogBuildListener extends RunListener<Run>
     payload.put("timestamp",
                 System.currentTimeMillis() / DatadogBuildListener.THOUSAND_LONG); // current time, s
     payload.put("status", status);
-    payload.put("tags", assembleTags(builddata, extraTags));
+    payload.put("tags", DataDogUtilities.assembleTags(builddata, extraTags));
 
     try {
-      post(payload, DatadogBuildListener.SERVICECHECK);
+      DataDogHttpRequests.post(payload, SERVICECHECK);
     } catch (Exception e) {
       logger.severe(e.toString());
     }
-  }
-
-  /**
-   * Sends a an event to the Datadog API, including the event payload.
-   *
-   * @param builddata - A JSONObject containing a builds metadata.
-   * @param extraTags - A list of tags, that are contributed through the {@link DataDogJobProperty}.
-   */
-  public final void event(final JSONObject builddata, HashMap<String,String> extraTags) {
-    logger.fine("Sending event");
-
-    // Gather data
-    JSONObject payload = new JSONObject();
-    String hostname = nullSafeGetString(builddata, "hostname");
-    String number = nullSafeGetString(builddata, "number");
-    String buildurl = nullSafeGetString(builddata, "buildurl");
-    String job = nullSafeGetString(builddata, "job");
-    long timestamp = builddata.getLong("timestamp");
-    String message = "";
-
-    // Setting source_type_name here, to allow modification based on type of event
-    payload.put("source_type_name", "jenkins");
-
-    // Build title
-    StringBuilder title = new StringBuilder();
-    title.append(job).append(" build #").append(number);
-    if ( "SUCCESS".equals( builddata.get("result") ) ) {
-      title.append(" succeeded");
-      payload.put("alert_type", "success");
-      message = "%%% \n [See results for build #" + number + "](" + buildurl + ") ";
-    } else if ( builddata.get("result") != null ) {
-      title.append(" failed");
-      payload.put("alert_type", "failure");
-      message = "%%% \n [See results for build #" + number + "](" + buildurl + ") ";
-    } else {
-      title.append(" started");
-      payload.put("alert_type", "info");
-      message = "%%% \n [Follow build #" + number + " progress](" + buildurl + ") ";
-    }
-    title.append(" on ").append(hostname);
-
-    // Add duration
-    if ( builddata.get("duration") != null ) {
-      message = message + durationToString(builddata.getDouble("duration"));
-    }
-
-    // Close markdown
-    message = message + " \n %%%";
-
-    // Build payload
-    payload.put("title", title.toString());
-    payload.put("text", message);
-    payload.put("date_happened", timestamp);
-    payload.put("event_type", builddata.get("event_type"));
-    payload.put("host", hostname);
-    payload.put("result", builddata.get("result"));
-    payload.put("tags", assembleTags(builddata, extraTags));
-    payload.put("aggregation_key", job); // Used for job name in event rollups
-
-    try {
-      post(payload, DatadogBuildListener.EVENT);
-    } catch (Exception e) {
-      logger.severe(e.toString());
-    }
-  }
-
-  /**
-   * Getter function to return either the saved hostname global configuration,
-   * or the hostname that is set in the Jenkins host itself. Returns null if no
-   * valid hostname is found.
-   * <p>
-   * Tries, in order:
-   *    Jenkins configuration
-   *    Jenkins hostname environment variable
-   *    Unix hostname via `/bin/hostname -f`
-   *    Localhost hostname
-   *
-   * @param envVars - An EnvVars object containing a set of environment variables.
-   * @return a human readable String for the hostname.
-   */
-  public final String getHostname(final EnvVars envVars) {
-    String[] UNIX_OS = {"mac", "linux", "freebsd", "sunos"};
-    String hostname = null;
-
-    // Check hostname configuration from Jenkins
-    hostname = getDescriptor().getHostname();
-    if ( (hostname != null) && isValidHostname(hostname) ) {
-      logger.fine(String.format("Using hostname set in 'Manage Plugins'. Hostname: %s", hostname));
-      return hostname;
-    }
-
-    // Check hostname using jenkins env variables
-    if ( envVars.get("HOSTNAME") != null ) {
-      hostname = envVars.get("HOSTNAME").toString();
-    }
-    if ( (hostname != null) && isValidHostname(hostname) ) {
-      logger.fine(String.format("Using hostname found in $HOSTNAME host environment variable. "
-                                + "Hostname: %s", hostname));
-      return hostname;
-    }
-
-    // Check OS specific unix commands
-    String os = getOS();
-    if ( Arrays.asList(UNIX_OS).contains(os) ) {
-      // Attempt to grab unix hostname
-      try {
-        String[] cmd = {"/bin/hostname", "-f"};
-        Process proc = Runtime.getRuntime().exec(cmd);
-        InputStream in = proc.getInputStream();
-        BufferedReader reader = new BufferedReader(new InputStreamReader(in));
-        StringBuilder out = new StringBuilder();
-        String line;
-        while ( (line = reader.readLine()) != null ) {
-          out.append(line);
-        }
-
-        hostname = out.toString();
-      } catch (Exception e) {
-        logger.severe(e.getMessage());
-      }
-
-      // Check hostname
-      if ( (hostname != null) && isValidHostname(hostname) ) {
-        logger.fine(String.format("Using unix hostname found via `/bin/hostname -f`. Hostname: %s",
-                                  hostname));
-        return hostname;
-      }
-    }
-
-    // Check localhost hostname
-    try {
-      hostname = Inet4Address.getLocalHost().getHostName().toString();
-    } catch (UnknownHostException e) {
-      logger.fine(String.format("Unknown hostname error received for localhost. Error: %s", e));
-    }
-    if ( (hostname != null) && isValidHostname(hostname) ) {
-      logger.fine(String.format("Using hostname found via "
-                                + "Inet4Address.getLocalHost().getHostName()."
-                                + " Hostname: %s", hostname));
-      return hostname;
-    }
-
-    // Never found the hostname
-    if ( (hostname == null) || "".equals(hostname) ) {
-      logger.warning("Unable to reliably determine host name. You can define one in "
-                     + "the 'Manage Plugins' section under the 'Datadog Plugin' section.");
-    }
-    return null;
-  }
-
-  /**
-   * Validator function to ensure that the hostname is valid. Also, fails on
-   * empty String.
-   *
-   * @param hostname - A String object containing the name of a host.
-   * @return a boolean representing the validity of the hostname
-   */
-  public static final Boolean isValidHostname(final String hostname) {
-    String[] localHosts = {"localhost", "localhost.localdomain",
-                           "localhost6.localdomain6", "ip6-localhost"};
-    String VALID_HOSTNAME_RFC_1123_PATTERN = "^(([a-zA-Z0-9]|"
-                                             + "[a-zA-Z0-9][a-zA-Z0-9\\-]*[a-zA-Z0-9])\\.)*"
-                                             + "([A-Za-z0-9]|"
-                                             + "[A-Za-z0-9][A-Za-z0-9\\-]*[A-Za-z0-9])$";
-    String host = hostname.toLowerCase();
-
-    // Check if hostname is local
-    if ( Arrays.asList(localHosts).contains(host) ) {
-      logger.fine(String.format("Hostname: %s is local", hostname));
-      return false;
-    }
-
-    // Ensure proper length
-    if ( hostname.length() > DatadogBuildListener.MAX_HOSTNAME_LEN ) {
-      logger.fine(String.format("Hostname: %s is too long (max length is %s characters)",
-                                hostname, DatadogBuildListener.MAX_HOSTNAME_LEN));
-      return false;
-    }
-
-    // Check compliance with RFC 1123
-    Pattern r = Pattern.compile(VALID_HOSTNAME_RFC_1123_PATTERN);
-    Matcher m = r.matcher(hostname);
-    if ( !m.find() ) {
-      return false;
-    }
-
-    // Passed all checks, so the hostname is valid
-    return true;
-  }
-
-  /**
-   * Returns an HTTP url connection given a url object. Supports jenkins configured proxy.
-   *
-   * @param url - a URL object containing the URL to open a connection to.
-   * @return a HttpURLConnection object.
-   * @throws IOException
-   */
-  public static HttpURLConnection getHttpURLConnection(final URL url) throws IOException {
-    HttpURLConnection conn = null;
-    ProxyConfiguration proxyConfig = Jenkins.getInstance().proxy;
-
-    if (proxyConfig != null) {
-      Proxy proxy = proxyConfig.createProxy(url.getHost());
-      if (proxy != null && proxy.type() == Proxy.Type.HTTP) {
-        logger.fine("Attempting to use the Jenkins proxy configuration");
-        conn = (HttpURLConnection) url.openConnection(proxy);
-        if (conn == null) {
-          logger.fine("Failed to use the Jenkins proxy configuration");
-        }
-      }
-    }
-
-    if (conn == null) {
-      conn = (HttpURLConnection) url.openConnection();
-      logger.fine("Using the Jenkins proxy configuration");
-    }
-
-    return conn;
-  }
-
-  /**
-   * Checks if a jobName is blacklisted, or not.
-   *
-   * @param jobName - A String containing the name of some job.
-   * @return a boolean to signify if the jobName is or is not blacklisted.
-   */
-  private final boolean isJobTracked(final String jobName) {
-    final String[] blacklist = blacklistStringtoArray( getDescriptor().getBlacklist() );
-    return (blacklist == null) || !Arrays.asList(blacklist).contains(jobName.toLowerCase());
-  }
-
-  /**
-   * Converts from a double to a human readable string, representing a time duration.
-   *
-   * @param duration - A Double with a duration in seconds.
-   * @return a human readable String representing a time duration.
-   */
-  public final String durationToString(final double duration) {
-    String output = "(";
-    String format = "%.2f";
-    if ( duration < DatadogBuildListener.MINUTE ) {
-      output = output + String.format(format, duration) + " secs)";
-    } else if ( (DatadogBuildListener.MINUTE <= duration)
-                && (duration < DatadogBuildListener.HOUR) ) {
-      output = output + String.format(format, duration / DatadogBuildListener.MINUTE)
-               + " mins)";
-    } else if ( DatadogBuildListener.HOUR <= duration ) {
-      output = output + String.format(format, duration / DatadogBuildListener.HOUR)
-               + " hrs)";
-    }
-
-    return output;
   }
 
   /**
@@ -693,41 +281,14 @@ public class DatadogBuildListener extends RunListener<Run>
   *
   * @return a String with a human-friendly OS name
   */
-  public final String getOS() {
+  public static String getOS() {
     String out = System.getProperty("os.name");
     String os = out.split(" ")[0];
 
     return os.toLowerCase();
   }
 
-  /**
-   * Safe getter function to make sure an exception is not reached.
-   *
-   * @param data - A JSONObject containing a set of key/value pairs.
-   * @param key - A String to be used to lookup a value in the JSONObject data.
-   * @return a String representing data.get(key), or "null" if it doesn't exist
-   */
-  public final String nullSafeGetString(final JSONObject data, final String key) {
-    if ( data.get(key) != null ) {
-      return data.get(key).toString();
-    } else {
-      return "null";
-    }
-  }
 
-  /**
-   * Converts a blacklist string into a String array.
-   *
-   * @param blacklist - A String containing a set of key/value pairs.
-   * @return a String array representing the job names to be blacklisted. Returns
-   *         empty string if blacklist is null.
-   */
-  public final String[] blacklistStringtoArray(final String blacklist) {
-    if ( blacklist != null ) {
-      return blacklist.split(",");
-    }
-    return ( new String[0] );
-  }
 
   /**
    * Getter function for the {@link DescriptorImpl} class.
@@ -781,8 +342,8 @@ public class DatadogBuildListener extends RunListener<Run>
 
       try {
         // Make request
-        conn = DatadogBuildListener.getHttpURLConnection(new URL(DatadogBuildListener.BASEURL
-                                                                 + DatadogBuildListener.VALIDATE
+        conn = DataDogHttpRequests.getHttpURLConnection(new URL(BASEURL
+                                                                 + VALIDATE
                                                                  + urlParameters));
         conn.setRequestMethod("GET");
 
@@ -829,7 +390,7 @@ public class DatadogBuildListener extends RunListener<Run>
      */
     public FormValidation doTestHostname(@QueryParameter("hostname") final String formHostname)
         throws IOException, ServletException {
-      if ( ( null != formHostname ) && DatadogBuildListener.isValidHostname(formHostname) ) {
+      if ( ( null != formHostname ) && DataDogUtilities.isValidHostname(formHostname) ) {
         return FormValidation.ok("Great! Your hostname is valid.");
       } else {
         return FormValidation.error("Your hostname is invalid, likely because"
