+++ src/main/resources/org/datadog/jenkins/plugins/datadog/DataDogJobProperty/config.jelly
@@ -1,20 +1,24 @@
 <j:jelly xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form">
-    <f:section title="DataDog Tagging"> 
+    <f:section title="DataDog Tagging">
         <f:optionalBlock name="enableFile" checked="${!instance.isTagFileEmpty()}" title="Add tags from file in workspace" inline="true">
             <f:entry title="File" field="tagFile">
                 <f:textbox/>
             </f:entry>
+            <f:entry title="Emit an extra event after checkout" field="emitOnCheckout"
+description="Note: Check this if you want to emit an extra event after checkout. Files are not immediately available when the build leaves the queue." >
+                <f:checkbox/>
+            </f:entry>
         </f:optionalBlock>
         <f:optionalBlock name="enableProperty" checked="${!instance.isTagPropertiesEmpty()}" title="Add tags from list of properties" inline="true">
-            <f:entry title="Properties" field="tagProperties">    
+            <f:entry title="Properties" field="tagProperties">
                 <f:textarea/>
-            </f:entry>        
-        </f:optionalBlock>                               
+            </f:entry>
+        </f:optionalBlock>
     </f:section>
     <tr>
         <td colspan="4">
             <hr style="background-color:#e5e5e5;height:1px;border:1px"/>
         </td>
-    </tr>   
+    </tr>
 </j:jelly>
